Add files via upload
Changed SurvivorModInfo.cs to provide a way to create passive skills (or grab an existing one from a character prefab). The skills do not actually have any functionality - my guess is that the game doesn't actually check if a character has the passive skill, but rather checks if the player is the right character. However, this could still be useful for explaining features that modders add to specific characters.